requirements.txt
djangorestframework-simplejwt
graphene==2.1.8
graphene-django==v2.14.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
Django==3.1.5
django-filter==2.4.0